Fix after pullRequest #37
