+++ BookStore/Modules/HomeModule/Helpful/Cell/CategoryCell.swift
@@ -1,5 +1,5 @@
 //
-//  CategoryCollectionViewCell.swift
+//  CategoryCell.swift
 //  BookStore
 //
 //  Created by Максим Горячкин on 05.12.2023.
@@ -8,7 +8,7 @@
 import UIKit
 
 class CategoryCell: UICollectionViewCell {
-    static let identifier = "CategoryCollectionViewCell"
+    static let identifier = String(describing: CategoryCell.self)
     
     lazy var categoryButton: UILabel = {
         let view = UILabel()
@@ -17,15 +17,13 @@ class CategoryCell: UICollectionViewCell {
         view.textAlignment = .center
         view.layer.borderColor = CGColor(red: 13/255, green: 13/255, blue: 13/255, alpha: 1)
         view.layer.borderWidth = 1
+        view.layer.cornerRadius = 6
         view.frame = contentView.bounds
         return view
     }()
     
     override init(frame: CGRect) {
         super.init(frame: frame)
-        
-//        contentView.clipsToBounds = true
-        contentView.layer.cornerRadius = 8
         contentView.addSubview(categoryButton)
     }
