+++ BookStore/Modules/HomeModule/Helpful/DataSources/BookSectionDataSource.swift
@@ -1,5 +1,5 @@
 //
-//  DataBuilder.swift
+//  BookSectionDataSource.swift
 //  BookStore
 //
 //  Created by Максим Горячкин on 05.12.2023.
