+++ BookStore/Modules/HomeModule/Model/HomeModel.swift
@@ -7,6 +7,8 @@
 
 import UIKit
 
+// MARK: - Book Model
+
 struct Book: Hashable, Identifiable {
     let id = UUID()
     let key: String
@@ -16,11 +18,15 @@ struct Book: Hashable, Identifiable {
     let imageID: Int
 }
 
+// MARK: - Categories
+
 enum HomeCategory: String, Hashable, CaseIterable {
     case week = "This Week"
     case month = "This Month"
     case year = "This Year"
 }
+
+// MARK: - Sections
 
 enum BookSection: Int, Hashable, CaseIterable {
     case categories
@@ -28,5 +34,19 @@ enum BookSection: Int, Hashable, CaseIterable {
     case recent
 }
 
+// MARK: - HomeViewModel
+
+struct HomeViewModel {
+    let seeAllTopBooksButton: Button
+    let topBooks: [Book]
+    let categories: [HomeCategory]
+    let seeAllRecentBooksButton: Button
+    let recentBooks: [Book]
+    
+    struct Button {
+        let title: String
+        let action: () -> Void
+    }
+}
